requirements.txt
ibllib>=1.5.13
ibl-pipeline-light>=0.0.0
statsmodels>=0.10.1
datajoint>=0.12
requests>=2.22.0
scikit_posthocs>=0.6.1
patsy>=0.5.1
seaborn<0.12
tqdm>=4.32.1
nose>=1.3.7
numpy>=1.16.4
pandas>=0.24.2
matplotlib>=3.0.3
lifelines>=0.25.0
pycircstat>=0.0.2
scikit-learn>=0.21.3
scipy>=1.3.0
